requirements.txt
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.2
pandas==0.25.1
psycopg2-binary==2.8.3
python-dateutil==2.8.0
pytz==2019.3
six==1.12.0
SQLAlchemy==1.3.9
Werkzeug==0.16.0
WTForms==2.2.1
